requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dotenv-python==0.0.1
frozenlist==1.3.3
grpcio==1.51.3
grpcio-tools==1.51.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
install==1.3.5
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
protobuf==4.22.1
pydantic==1.10.7
python-dotenv==1.0.0
pytz==2022.7.1
qdrant-client==1.1.0
requests==2.28.2
rfc3986==1.5.0
slack-bolt==1.18.0
slack-sdk==3.21.3
sniffio==1.3.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2